Better invocation of bash, BSD workarounds
Using `/bin/bash` to invoke bash is effective, but does not allow for
flexability. `/usr/bin/env bash` will load bash from the PATH. This
allows for the user to set the version of bash.

As for the BSD replacements, I basically detect the presence of GNU
versions of the coreutils, and if they are present, create functions
to use them in place of the BSD versions

Helps fix issue #5